flow [nfc]: Convert $ElementType to indexed-access syntax
For the same reasons as for $PropertyType in the preceding commit.

Done mostly with another crude search-and-replace:

    $ perl -i -0pe '
          s/\$ElementType<(.*?),\s*(.*?)>/($1)[$2]/gs
        ' src/**/*.js

Then manually fixed the only place that got wrong, plus the spot in
types/react-intl.js.flow it would have gotten wrong.

Leave in place a number of uses of $ElementType in TsFlower output.
The right way to update those will be for TsFlower to switch to
emitting the new syntax.