example data: Fix date_joined on users to always be valid
This was coming out as "2014-04-00" 1/30th of the time, which isn't
a valid date string.

As a result, when we go to parse that string in
getHasUserPassedWaitingPeriod or in its tests, it comes out as NaN
and things go haywire.  This caused the flake #5404.

Fix by making the day run from 1 through 30, instead of 0 through 29.

Fixes: #5404