exampleData: Reserve some message, stream, and user IDs for snapshot tests
In generateInboundEventEditSequence-test.js, when we build state
objects from eg.plusReduxState, the data we add includes some
constant message, stream, and user IDs. Constant because they end up
appearing in test snapshots, and we want those snapshots to be
stable.

This is a problem when the data we're building from already claims
those IDs, for example, by saying that a stream with ID 1 exists and
that we have a subscription for it. Since eg.plusReduxState uses
random IDs, this has been happening some small fraction of the time,
resulting in flakes where the snapshot says a stream is *not*
subscribed, but the input data says it *is*. That's #5415.

So, to fix, reserve a small section of stream IDs, including 1 and
2, which we use in the mentioned snapshot tests, so that
eg.plusReduxState never chooses them. And similarly reserve a range
of message and user IDs, too. As suggested by Greg:
  https://github.com/zulip/zulip-mobile/pull/5415#discussion_r898260623

In another approach to fix those flakes, we might have instead built
on eg.plusReduxState by erasing some of it first, removing spreads
like the one in

  streams: [...eg.plusReduxState.streams, stream1, stream2],

But as we mentioned in 371e9b48e, that'd risk corrupting the state
we get from eg.plusReduxState. That's the only example state that
claims to offer a realistic uncorrupted state with server data, and
it's good not to interfere with that guarantee if we can avoid it.
(For example, if eg.plusReduxState grows a reference to a stream in
`muteModel`, it'd be corrupt if that stream is absent in the
`streams` state.) It's too bad that the easy, natural thing to do is
to omit the spreads.

Fixes: #5415