SwitchRow [nfc]: Remove unused `style` prop
No callers currently use this. Might as well remove it, because it
could invite excessive micromanaging of individual style attributes
by callers.

If SwitchRow wants to offer more customization in the future, it can
do so with purpose-designed props that don't have that problem. (For
an example of this approach, see the leftMargin and rightMargin
props of ZulipTextButton.)