UserStatusScreen [nfc]: Use codeToEmojiMap, not parseUnicodeEmojiCode
And unexport parseUnicodeEmojiCode, since this was its only use
outside the module. Give it a jsdoc, too, since it doesn't have one
yet.

This wouldn't be NFC if any of statusSuggestions were either
represented in `overrides` (none of them are) or not represented in
codeToEmojiMap (they all are).

Now, this site follows the pattern of all the other places where
we're holding an `emoji_code` and we want the actual Unicode emoji
character. Soon, we'll refactor all of these sites so that they can
consult server data, when available, for #2956.