emoji: Use ServerEmojiData, when present, to display Unicode emoji
For where we treat each Unicode emoji in the WebView CSS
(src/webview/css/cssEmojis.js), this means mapping over the emoji
codes in ServerEmojiData, if present, and otherwise using the
current hard-coded data as a fallback.

For where we render Unicode emoji outside the WebView, this means
having displayCharacterForUnicodeEmojiCode take the ServerEmojiData,
if present, and otherwise fall back to the current hard-coded data.
(Most of the changes in this commit are to pipe the ServerEmojiData
from Redux down to displayCharacterForUnicodeEmojiCode.)