internalLinks [nfc]: Early-return getNarrowFromLink on 'non-narrow' type
Soon we'd like to make it official that getPathsFromUrl expects a
url that appears to be a narrow link, i.e., one where
`isNarrowLink(url, realm)` is true. This early return, before the
getPathsFromUrl call, makes sure we meet that expectation.

This is NFC because before this, in the 'non-narrow' case, `paths`
was always getting set and then ignored.

While we're at it, make the `switch` exhaustive with
ensureUnreachable.