internalLinks: Simplify how we ignore #narrow segment in getHashSegments[…]
This isn't quite NFC, but it could only make a difference in the
presence of a server bug.

Locally, we have an invariant that `new URL(url, realm).hash` starts
with "#narrow/". That's guaranteed by the fact that
`isNarrowLink(url, realm)` is true, as specified in the jsdoc.

So as long as "#narrow/" only occurs that one time in the hash, this
is NFC. That's definitely the expected case.

But our client-side code doesn't enforce that. If somehow the server
failed to URL-encode the hash, and we got
  #narrow/stream/7-test-here/topic/#narrow/
instead of
  #narrow/stream/7-test-here/topic/.23narrow.2F
, then the old code would've initialized `result` to
  ['']
instead of
  ['stream', '7-test-here', 'topic', '#narrow/']
. The latter seems more natural to do here, even though the lack of
encoding would be unexpected and could lead to problems later on.

Anyway, I don't think such a server bug has been active, but I'm not
marking this commit NFC, just in case.