flags: Cut never-used flags from state.flags
These all used to be genuine UserMessage flags in the server's
model, and so presumably in the API, but no longer are (and
haven't been for years.)

In zulip-mobile, they were all first mentioned in 72f33910a in
2017-06, and they all have the same history here: we've never
touched them except in (a) initializing the flags state and
(b) tests.

The name `is_me_message` shows up a bit more in the history because
it's also the name of a property on Message itself, but its history
as a flag is the same.  (For that matter, we've never actually used
the Message property either.)