LightboxScreen: Require URL object, not string, for props.route.params.src
This means that we won't open the lightbox at all if the image's src
isn't accepted by the URL parser. Contrast the behavior just before
this commit, where we would let the lightbox open, then close it
immediately (via navigation.goBack()) upon finding out that the src
doesn't parse.

We might've put a warning on `handleImage` about the `new URL` call
adding some computation time. But it's just not the kind of function
you would call more than once per user action, and also, a fraction
of a millisecond won't make a difference when the function's job is
to open the lightbox, which involves a nav animation.