api.getFileTemporaryUrl: Stop requesting with double slash, as in api/v1//…
CZO today seems happy to accept the double slash, returning the same
thing it returns when we don't send the double slash. Still, better
not to send it, since it's not what zulip.yaml asks for.

We also drop any query or fragment that might have been in `href`.
As Greg points out at
  https://github.com/zulip/zulip-mobile/pull/5492#discussion_r1008536997 :
> That's also a good change -- a query could potentially pass
> unwanted extra parameters to the API endpoint we're talking to,
> and a URL fragment isn't meant to go to a server at all.