offline notice: Move to always be at top of screen, just below top inset
We want to try making the OfflineNotice a surface that fills the top
inset, instead of being a bar that cuts through the middle of an
existing surface.

Toward that goal, make it so OfflineNotice abuts the top inset, to
simplify the diff when it starts actually covering the top inset.

In some cases, the OfflineNotice is already where we want it to be.

In all cases, take care to make sure the OfflineNotice isn't cut off
before it reaches the the extreme left and right of the screen. In
ModalNavBar and ModalSearchNavBar, we do so by moving two sets of
horizontal padding inward to the view that's styled
`styles.contentArea`, and putting the OfflineNotice outside that
view: automatic inset padding from a SafeAreaView, and our own added
padding of 4px. Having done that, remove an old comment whose
explanation would now be wrong about why we have the inner view.