android build: Upgrade Gradle to 7.3
Done to follow the template-app change in
facebook/react-native@c180627ac, on the path to the RN v0.68
upgrade.

Done by following the "How to upgrade Gradle" instructions at the
top of gradle-wrapper.properties. The template-app commit makes a
bunch of changes that look like the ones Gradle wanted to make in
the 7.2 bump, so it's probably just belatedly catching up with
those. See previous commit.