deps: Bump react-native-photo-view to (still) unreleased commit, for AGP 7
Before this change, we were pointing Yarn to a commit from 2018 on
this repo's `master` branch, since a new release with the commit
hadn't been made to NPM in a reasonable time. (It still hasn't.)

Now, just bump to this project's current `master`, which is three
commits ahead of that commit:
  https://github.com/alwx/react-native-photo-view/commits/master

In particular, we want alwx/react-native-photo-view@ef532a348, which
has the project's build.gradle stop using the `compile` dependency
configuration in favor of `implementation`. The former has long been
deprecated, and is an error with AGP 7, which we'd like to use soon:
  https://developer.android.com/studio/releases/gradle-plugin#dependency-configurations-removed

We'd still like to stop using this library since it's almost totally
unmaintained; that's #4217.

Anyway, I tested on my iPhone 13 Pro running iOS 15.6, and on the
office Android device (Samsung Galaxy S9 running Android 9), and I
didn't notice any change in behavior of the lightbox.

Regarding the `jcenter` exception for `photodraweeview` (see
android/build.gradle in this commit), I tried removing the exception
entirely, hoping that 1.1.3 might be available in the non-jcenter
repos consulted by the build, even though 1.0.0 was not. But no, I
again got an error that `photodraweeview` could not be found, so it
looks like we still need to make an exception. See discussion of
some further findings:
  https://github.com/zulip/zulip-mobile/pull/5507#discussion_r993965873