deps: Upgrade to Expo SDK 45
Following the "Bare workflow" instructions for upgrading:
  https://blog.expo.dev/expo-sdk-45-f4e332954a68#5954

As with the last one, 44 in 3edc37ae4, the `expo upgrade` command
went through a phase it described as "Updating packages to
compatible versions (where known)," and that changed several of our
dependency ranges, even of non-Expo things. (For speculation on what
that phase is about, see 3edc37ae4.)

We rejected version-range changes to these libraries that would have
resulted in downgrades to the libraries' resolved versions:

- react-native-safe-area-context from 4.3.1 to 4.2.4
- react-native-screens from 3.13.1 to 3.11.1
- react-native-webview from 11.22.2 to 11.18.1

We rejected upgrades to @react-native-community/netinfo and
react-native-reanimated for the reasons we gave in the Expo 44
upgrade; see those at
  https://github.com/zulip/zulip-mobile/pull/5441#issuecomment-1184944562 .

And we rejected bumping react-native from 0.67.4 to 0.68.2, with
optimism that we can get this Expo upgrade done before that RN
upgrade, which is #5344.

We took the rest.

For the step

> Update your `App.Delegate.mm` (you should have moved from .m to
> .mm in the previous step) according to this diff

, we found that we'd already taken some of those changes when Expo
backported them to the Expo 44 template app. Others, like the .m to
.mm rename, we'd like to postpone until we do the RN 68 upgrade. A
full audit of Expo's template-app commits from 44 to 45 found no
changes we'd be interested in applying, except some related to RN 68
that we'd like to do with that upgrade; I've mentioned those on our
RN 68 upgrade issue. See:
  https://github.com/zulip/zulip-mobile/pull/5507#issuecomment-1324181780

I tested basic functionality on my iPhone 13 Pro running iOS 15.6,
and on the office Android device, a Samsung Galaxy S9 running
Android 9. I didn't see any problems with building or running.