msglist [nfc]: Make SinglePageWebView a real React component
From some experimentation, it turns out that re-rendering a WebView
element is perfectly fine, just as one would hope if WebView is to
be a well-behaved React component type.

What isn't fine is re-rendering it with a *different `html` value*.
Doing that will cause the page to be reloaded, so that any UI state
inside it gets reset -- most notably, the user gets scrolled right
back to their starting point (i.e. the first unread, or else the
very end).  Which is pretty reasonable on the WebView's part.

So that means that the way our current logic succeeds at avoiding
that problem isn't that it avoids re-rendering the WebView (though
it does do that); it's that it avoids re-rendering the MessageList,
and thereby avoids ever computing a new `html` in the first place.

As a result, it's perfectly fine to have another component
interposed here.  Even if this new component gets re-rendered for
some arbitrary reason, it'll use the `html` value it got from its
parent, namely MessageList -- which we continue to take care not to
re-render, so the `html` value still never changes.