react [nfc]: Fix a reference to "legacy context API"
The API this is referring to is actually the modern context API,
just the class-component form of it.

The "legacy context API" is something else, involving a static
property `contextTypes`.  We haven't used that one for anything
since commit 30e4d19cc, PR #4199, back in 2020-07.

Also mention what's now the most common alternative to withGetText:
the useContext hook.