theme: Provide name in ThemeData; switch to context in MessageList
Change the way `MessageList` updates its theme. It now gets
its ThemeData information from the `ThemeContext` instead of
its selector props.

This is helpful if the theme ever changes, by causing the component
to re-render rather than get the colors out of sync.  That currently
isn't possible while a `MessageList` is mounted, because the theme
can only change by navigating to the settings screen; but it will
become possible with #4009 via the OS-level theme changing.