autocomplete types: Remove AutocompleteOption type; simply use UserOrBot
This type saved us the trouble of faking every last boring detail of
a whole UserOrBot object, in the world where we used fake user-like
objects for wildcard mentions. We stopped using those in the
previous commit, so we don't need this type anymore, and we can just
say we're passing around whole UserOrBot objects because that's what
we're doing.