deps: Upgrade `@react-native-camera-roll/camera-roll` to 5.x
This also means following a rename, from the former
`@react-native-community/cameraroll`.

We had had our own small patch to this library, making the Flow
types in its implementation compatible with modern Flow practice:
  https://github.com/react-native-cameraroll/react-native-cameraroll/pull/328
In v5, that became obsolete because they converted the implementation
from Flow to TypeScript.  So just drop the patch.

That also means we're losing type-checking on the library's interface.
We'll get types back with TsFlower in the next commit; for now just
leave a fixme.  We use very little API surface area here, so it's not
essential to have types on it.