SearchMessagesScreen [nfc]: Pass scrollEnabled={false} to <Screen />
The default value for scrollEnabled is true, which doesn't sound
like what we want applied in this case. When there is anything to
scroll through, it'll be search results, and when we're showing
search results, we want the WebView to handle scrolling.

A `true` value wasn't doing anything, thank goodness, because the
height of the content area (containing a loading or no-results
indicator, or a WebView with results) was always set to match the
height of the ScrollView itself, because of:

- a `flex: 1` in the outermost view rendered by SearchMessagesCard,
  and
- a `flex: 1` passed to the ScrollView's contentContainerStyle.

Still, best for SearchMessagesScreen to tell the Screen component
explicitly that the Screen shouldn't try to manage scrolling.