Screen: Move padding into ScrollView's content area
Before, the vertical padding decreased the height of the window onto
the scrollable area.

Now, the vertical padding is *in* the scrollable area, and it can be
scrolled offscreen, giving more room for the actual content to show.
This is especially helpful on a phone in in landscape mode, where
vertical space is tight.

Affected screens are the ones where true is passed for `padding` and
`scrollEnabled` is true (or unset, because it defaults to true):

  AccountPickScreen
  AuthScreen
  PasswordAuthScreen
  RealmInputScreen
  CreateStreamScreen
  EditStreamScreen