accountsReducer: Invariant that ACCOUNT_REMOVE's index is in bounds
If `action.index` is negative, the behavior of `.splice` seems (from
an experiment) to be to count from the end of the array, so
`.splice(-1, 1)` would cut the last element off. This would cause
confusion if it ever happened, i.e., if the one ACCOUNT_REMOVE
dispatcher, in AccountPickScreen, passed a negative index.
Fortunately, it doesn't. The index comes from React Native's
`FlatList`: it's the index for one of the account items to render. I
don't think React Native would ever make that negative.

If `action.index` is too high for the array, then (experimentally)
`.splice` would leave it alone, which is better than corrupting the
data. In theory, this could happen: somehow `state.accounts` changes
after the "Remove account" confirmation dialog appears, such that
when the user presses "Confirm", the index is out of bounds. We
don't want this to happen (hence the invariant), but,
experimentally, `.splice()` would just leave the accounts state
unchanged, which is better than changing it in a weird way.