account actions [nfc]: Make loginSuccess thunk action a GlobalThunkAction
Just as much as accountSwitch, this thunk action does touch the
global `state.accounts`, by either adding a new account at the front
of the array, or -- just like accountSwitch -- moving an existing
account to the front.

So, make it so callers treat it as a GlobalThunkAction, and have it
act like one in its implementation.