flags model: Fix bug where mark-all-as-read would *add* unread markers
(Edit: Actually, the previous commit changed the symptom of #5597,
making this commit message a bit inaccurate. At this point, things
have changed so that "Mark as read" won't add unread markers, it'll
just fail to remove them. See discussion on the PR.)

You're not meant to use Object.keys on an Immutable.Map; details at
issue #5597 for how this caused a display-only bug where messages in
view would get marked with unread markers.

Greg points out, at
  https://github.com/zulip/zulip-mobile/issues/5597#issuecomment-1344624167

> If we'd had type-checking on those tests, Flow would have pointed
> out that the test data needed to change when we changed the
> definition of `MessagesState`, which would have alerted us that
> `flagsReducer` was consuming data of that type and given us a
> better chance to see we needed to update that logic.

We'll start type-checking flagsReducer-test.js soon. But for now,
just for this one test, give input that's unchecked but well formed,
and fix the reducer's implementation. The fixed test passes with
this new reducer implementation, and it would fail with the old,
buggy implementation.

Fixes: #5597