presence tests: Fill in other `ClientPresence`s next to `aggregated`s
Every UserPresence is supposed to have an `aggregated` key, and soon
we'll bring Flow into this file to help us make sure we include it.

But UserPresence has another requirement, and I'm not sure if
there's a nice way to get Flow to help with it: it's *also* supposed
to have one or more other `ClientPresence`s that `aggregated` is
aggregated across.

So, in this separate commit before bringing in Flow, fill in a
`ClientPresence` next to each `aggregated` that doesn't have one.