ZulipText [nfc]: Offer inherit{FontSize,Color} props
This is a new approach to dealing with ZulipText's
aggressiveDefaultStyles. Like the current one, it has benefits and
tradeoffs.

Here's an example of an upside: in RealmInputScreen, we can now
group three ZulipText(Intl)s together under one ZulipText that
provides `styles.suggestionText` through inheritance, without having
to repeat the very specific `fontSize: 12` in some form across all
three of them. We do have to pass a new prop, inheritFontSize, but
that actually feels appropriate here: those three text components
are overriding a global default, after all, and now they all have
breadcrumbs leading back to the one element that sets the overriding
value.

Oh, and another nice thing: WebLink's implementation is much less
finicky now! I'm not aware of any bugs it has caused, but it feels
easier to understand, and more resilient to upgrading React across
breaking changes.

A downside is that WebLink's children now have to know that
WebLink's set of provided styles includes a `color` and do a little
incantation (`inheritColor`) to let that work. It seems natural for
WebLink's callers to be aware of that, but the incantation could get
tiresome if WebLink (or some analogous component) wanted to apply a
bunch of other attributes, like a text underline, etc.