ios: Require iOS 14.0+
Only 0.7% of our users are on 13 and older; see
  https://chat.zulip.org/#narrow/stream/48-mobile/topic/platform.20versions/near/1476154

In fact, only 3.8% of our iOS users are on iOS 14, so graceful
degradation below iOS 15 is fine.

The bump to iOS 14 in tools/generate-webview-js seems to cause
@babel/preset-env to drop @babel/plugin-transform-template-literals,
so add an `include` key to force it to be included, with a note
about why. See the doc on `include`:
  https://babeljs.io/docs/en/babel-preset-env#include
and the doc on @babel/plugin-transform-template-literals:
  https://babeljs.io/docs/en/babel-plugin-transform-template-literals#loose