tools/test: Simplify handling of suites list
This makes for a bit less duplication in the code; fewer places
to have to update when adding a new suite.  The CLI-parsing logic
also gets a bit simpler.

I think this also makes the usage message a bit simpler to understand.
Like the existing version, it's a bit awkward as long as there are no
non-default suites; but we'll be adding one of those shortly.