DiagnosticsScreen: Remove, putting app version directly on settings screen
We're about to add a per-account troubleshooting screen for
notifications, and link to it from a better location: the
notification settings screen. Best to remove this screen so that
people don't get distracted by it when they want to troubleshoot
notifications.

Except for the app version, I can't remember anything on this screen
or its nested screens being useful. Also, we've had at least one
request to make the app version number more prominent, which this
does. See discussion:
  https://github.com/zulip/zulip-mobile/pull/5654#discussion_r1094955584