ProfileScreen: Put status and set-your-status link in the same row
Before this, we were separating the line that showed the user's
status from the line that led to the UI to change the status (a
button linking to UserStatusScreen) -- with the "Invisible mode"
switch in between.

Now, it's a single NestedNavRow that shows the emoji and status
text, if any. Tapping it leads to the UserStatusScreen.

Also reword the user-status screen's title a bit to avoid the
awkward user-facing word "user":
  https://github.com/zulip/zulip-mobile/pull/5665#issuecomment-1435404059