ios: Fix bug where tapping a notification didn't navigate to conversation
For no apparent reason, the library we've been using for iOS
notifications (@react-native-community/push-notification-ios) sends
remote notification response events, including the simple response
of tapping on a remote notification, with the name
'localNotification' instead of the name they document, which is
'notification':
  https://github.com/react-native-push-notification/ios/blob/v1.10.1/README.md#how-to-determine-push-notification-user-click

(For what their 'notificaton' event actually means, see a recent
commit where we removed our handler for it.)

Rather than giving in and just listening for 'localNotification',
set up our own native module to forward the event that iOS sends to
the app. This took some time, but it makes the logic more
transparent and removes some dependence on that library, which are
nice outcomes.

There was some custom `RCTConvert` and JavaScript munging to read
through in the library's codepath, but thankfully the part that
affects what we actually use is small: just a call to
`RCTJSONClean`.

Fixes: #5679