ZLPNotifications [nfc]: Remove trivial `constantsToExport`
Since the module doesn't have any constants to export. If it wants
to export any, we can always add the method back.

Also remove `requiresMainQueueSetup`, which (according to the linked
doc) was needed because we were implementing `constantsToExport`.

On my iPhone 13 Pro running iOS 16.1, the feature supported by this
module (alerting the user when Zulip notifications are disabled at
the system level) still works as expected.