unread [nfc]: Cut always-false isMuted properties
These were always false in the data structure actually returned
to the UI; their presence there was basically a leak of an
implementation detail, as the purpose they served was in the
communication between getUnreadStreamsAndTopics and its helper.
Now that we've erased that boundary, we don't need them at all.