mute [nfc]: Rename getter to isTopicVisibleInStream, and flip sense
With the upcoming feature of unmuted topics in muted streams,
the question "is this topic muted?" will split into several
different questions that can have different answers.
To prepare for that, stop having a function asking that
less-specific question.

For several call sites, we can cleanly write a name and jsdoc
for a function that covers all of them.

In the topic action sheet, the question we'll end up wanting to ask
is different from those, and reflects a UI choice we're making
specifically about the action sheet.  So just put an explicit
switch statement there.